requirements-dev.txt
asgiref==3.11.1
boto3==1.42.89
botocore==1.42.89
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
dj-database-url==3.1.2
Django==5.2.12
django-cors-headers==4.9.0
django-storages==1.14.6
djangorestframework==3.17.1
drf-yasg==1.21.15
easyocr==1.7.2
filelock==3.29.0
fsspec==2026.3.0
gunicorn==25.3.0
idna==3.11
ImageIO==2.37.3
inflection==0.5.1
Jinja2==3.1.6
jmespath==1.1.0
lazy-loader==0.5
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
ninja==1.13.0
numpy==2.4.4
opencv-python-headless==4.13.0.92
packaging==26.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.2.0
psycopg2-binary==2.9.11
pyclipper==1.4.0
pycparser==3.0
PyJWT==2.12.1
pypdfium2==5.7.1
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.33.1
resend==2.28.1
s3transfer==0.16.0
scikit-image==0.26.0
scipy==1.17.1
shapely==2.1.2
six==1.17.0
sqlparse==0.5.5
sympy==1.14.0
tifffile==2026.3.3
torch==2.11.0
torchvision==0.26.0
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
whitenoise==6.12.0
